+++ src/cli/idm/idm-export.ts
@@ -22,7 +22,12 @@ export default function setup() {
         'Config entity name. E.g. "managed", "sync", "provisioner-<connector-name>", etc.'
       )
     )
-    .addOption(new Option('-f, --file [file]', 'Export file. Ignored with -A.'))
+    .addOption(
+      new Option(
+        '-f, --file [file]',
+        'Export file (or directory name if exporting mappings separately). Ignored with -A.'
+      )
+    )
     .addOption(
       new Option(
         '-E, --entities-file [entities-file]',
@@ -47,6 +52,12 @@ export default function setup() {
         'Export all IDM configuration objects into separate JSON files in directory -D. Ignored with -N, and -a.'
       )
     )
+    .addOption(
+      new Option(
+        '-s, --separate-mappings',
+        'Export sync.json mappings separately in their own directory.'
+      )
+    )
     .action(
       // implement command logic inside action handler
       async (host, realm, user, password, options, command) => {
@@ -61,7 +72,11 @@ export default function setup() {
         // export by id/name
         if (options.name && (await getTokens())) {
           verboseMessage(`Exporting object "${options.name}"...`);
-          const outcome = await exportConfigEntity(options.name, options.file);
+          const outcome = await exportConfigEntity(
+            options.name,
+            options.file,
+            options.separateMappings
+          );
           if (!outcome) process.exitCode = 1;
         }
         // require --directory -D for all-separate functions
@@ -89,7 +104,8 @@ export default function setup() {
           );
           const outcome = await exportAllConfigEntities(
             options.entitiesFile,
-            options.envFile
+            options.envFile,
+            options.separateMappings
           );
           if (!outcome) process.exitCode = 1;
           await warnAboutOfflineConnectorServers();
@@ -99,7 +115,9 @@ export default function setup() {
           verboseMessage(
             `Exporting all IDM configuration objects into separate files in ${state.getDirectory()}...`
           );
-          const outcome = await exportAllRawConfigEntities();
+          const outcome = await exportAllRawConfigEntities(
+            options.separateMappings
+          );
           if (!outcome) process.exitCode = 1;
           await warnAboutOfflineConnectorServers();
         }