+++ src/cli/mapping/mapping-list.ts
@@ -1,0 +1,39 @@
+import { Option } from 'commander';
+
+import { getTokens } from '../../ops/AuthenticateOps';
+import { listMappings } from '../../ops/MappingOps';
+import { verboseMessage } from '../../utils/Console';
+import { FrodoCommand } from '../FrodoCommand';
+
+export default function setup() {
+  const program = new FrodoCommand('frodo mapping list');
+
+  program
+    .description('List IDM mappings.')
+    .addOption(
+      new Option('-l, --long', 'Long with all fields.').default(false, 'false')
+    )
+    .action(
+      // implement command logic inside action handler
+      async (host, realm, user, password, options, command) => {
+        command.handleDefaultArgsAndOpts(
+          host,
+          realm,
+          user,
+          password,
+          options,
+          command
+        );
+        if (await getTokens()) {
+          verboseMessage(`Listing all mappings`);
+          const outcome = await listMappings(options.long);
+          if (!outcome) process.exitCode = 1;
+        } else {
+          process.exitCode = 1;
+        }
+      }
+      // end command logic inside action handler
+    );
+
+  return program;
+}