+++ test/e2e/idm-export.e2e.test.js
@@ -50,9 +50,10 @@
 FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export --name script
 FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export -N script -f test.json
 FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export -N script -D testDir4
-FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export -AD testDir1
+FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export -sN sync
+FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export -sAD testDir1
 FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export -AD testDir2 -E test/e2e/env/testEntitiesFile.json -e test/e2e/env/testEnvFile.env
-FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export --all-separate --directory testDir3 --entities-file test/e2e/env/testEntitiesFile.json --env-file test/e2e/env/testEnvFile.env
+FRODO_MOCK=record FRODO_NO_CACHE=1 FRODO_HOST=https://openam-frodo-dev.forgeblocks.com/am frodo idm export --all-separate --separate-mappings --directory testDir3 --entities-file test/e2e/env/testEntitiesFile.json --env-file test/e2e/env/testEnvFile.env
 */
 import { testExport } from './utils/TestUtils';
 import { connection as c } from './utils/TestConfig';
@@ -87,9 +88,15 @@ describe('frodo idm export', () => {
     await testExport(CMD, env, type, undefined, dirName, false);
   });
 
-  test('"frodo idm export -AD testDir1": should export all idm config entities to separate files in the "testDir" directory', async () => {
+  test('"frodo idm export -sN sync": should export the idm config entity with idm id "sync" separately', async () => {
+    const dirName = 'sync';
+    const CMD = `frodo idm export -sN sync`;
+    await testExport(CMD, env, 'sync', undefined, dirName, false);
+  });
+
+  test('"frodo idm export -sAD testDir1": should export all idm config entities to separate files in the "testDir" directory', async () => {
     const dirName = 'testDir1';
-    const CMD = `frodo idm export -AD ${dirName}`;
+    const CMD = `frodo idm export -sAD ${dirName}`;
     await testExport(CMD, env, undefined, undefined, dirName, false);
   });
 
@@ -99,9 +106,9 @@ describe('frodo idm export', () => {
     await testExport(CMD, env, undefined, undefined, dirName, false);
   });
 
-  test('"frodo idm export --all-separate --directory testDir3 --entities-file test/e2e/env/testEntityFile.json --env-file test/e2e/env/testEnvFile.env": should export all idm config entities to separate files in the "testDir" directory according to the entity and env files', async () => {
+  test('"frodo idm export --all-separate --separate-mappings --directory testDir3 --entities-file test/e2e/env/testEntityFile.json --env-file test/e2e/env/testEnvFile.env": should export all idm config entities to separate files in the "testDir" directory according to the entity and env files', async () => {
     const dirName = 'testDir3';
-    const CMD = `frodo idm export --all-separate --directory ${dirName} --entities-file test/e2e/env/testEntitiesFile.json --env-file test/e2e/env/testEnvFile.env`;
+    const CMD = `frodo idm export --all-separate --separate-mappings --directory ${dirName} --entities-file test/e2e/env/testEntitiesFile.json --env-file test/e2e/env/testEnvFile.env`;
     await testExport(CMD, env, undefined, undefined, dirName, false);
   });
 });